moved basicAuthPlugin to auth folder
