+++ src/components/context/ContextProvider.tsx
@@ -16,7 +16,7 @@ import { Store } from "redux";
 import DataFetcher from "opds-web-client/lib/DataFetcher";
 import ActionsCreator from "opds-web-client/lib/actions";
 import { adapter } from "opds-web-client/lib/OPDSDataAdapter";
-import basicAuthPlugin from "../../authPlugin";
+import basicAuthPlugin from "../../auth/basicAuthPlugin";
 
 type ProviderProps = PreloadedData & {
   // we allow custom store and actions
